Refactoring and adding receptionReport.clientId
