+++ app/src/main/java/com/fivegmag/a5gmsmediastreamhandler/player/QoeMetricsReporter.kt
@@ -8,7 +8,9 @@ interface QoeMetricsReporter {
 
     fun getQoeMetricsReportingScheme() : String
 
-    fun getQoeMetricsReport(qoeMetricsRequest: QoeMetricsRequest) : String
+    fun getQoeMetricsReport(qoeMetricsRequest: QoeMetricsRequest, reportingClientId: String) : String
 
     fun reset()
+
+    fun resetState()
 }
